+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 - Fixed \\ import issue
 - Added graceful handling for when expected json sections are not found
 - Reorganized repo structure (src folder)
+- Added ability to add answers and answer fields
+- Added ability to delete answers